combine more tests
