+++ server/service/certificate_test.go
@@ -12,6 +12,10 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+////////////////////////////////////////////////////////////////////////////////
+// Service
+////////////////////////////////////////////////////////////////////////////////
+
 func TestCertificateChain(t *testing.T) {
 	server, teardown := setupCertificateChain(t)
 	defer teardown()
